Updated LOVD_Object to handle linking tables
- Updated LOVD_Object::loadEntry(), LOVD_Object::viewEntry() and
LOVD_Object::getCount() to allow multiple keys to be passed to them.
- Removed redundant code and used the updated loadEntry and viewEntry
for gene panels.
- Changed the transcriptid column in the gene panel tables to mediumint.

+++ src/install/inc-sql-tables.php
@@ -859,7 +859,7 @@ $aTableSQL =
     'CREATE TABLE ' . TABLE_GP2GENE . ' (
     genepanelid SMALLINT(5) UNSIGNED ZEROFILL NOT NULL,
     geneid VARCHAR(25) NOT NULL,
-    transcriptid SMALLINT(5) UNSIGNED ZEROFILL,
+    transcriptid MEDIUMINT(8) UNSIGNED ZEROFILL,
     inheritance VARCHAR(50),
     pmid INT(10) UNSIGNED,
     remarks TEXT,
@@ -883,7 +883,7 @@ $aTableSQL =
     'CREATE TABLE ' . TABLE_GP2GENE_REV . ' (
     genepanelid SMALLINT(5) UNSIGNED ZEROFILL NOT NULL,
     geneid VARCHAR(25) NOT NULL,
-    transcriptid SMALLINT(5) UNSIGNED ZEROFILL,
+    transcriptid MEDIUMINT(8) UNSIGNED ZEROFILL,
     inheritance VARCHAR(50),
     pmid INT(10) UNSIGNED,
     remarks TEXT,
